pyproject.toml
[project]
name = "test-task-medol"
version = "0.1.0"
description = ""
authors = [{ name = "Borovkov Ilya", email = "borovkovkin@gmail.com" }]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "fastapi[standart] (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.1,<0.35.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "sqlalchemy[asyncio] (>=2.0.40,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.6"
